+++ src/irmin-fs/irmin_fs.mli
@@ -68,8 +68,8 @@ end
 
 module Append_only (IO : IO) : Irmin.Append_only.Maker
 module Atomic_write (IO : IO) : Irmin.Atomic_write.Maker
-module Maker (IO : IO) : Irmin.Maker
-module KV (IO : IO) : Irmin.KV_maker
+module Maker (IO : IO) : Irmin.Maker with type info = Irmin.Info.default
+module KV (IO : IO) : Irmin.KV_maker with type info = Irmin.Info.default
 
 (** {2 Advanced configuration} *)
 
@@ -88,7 +88,9 @@ end
 
 module Append_only_ext (IO : IO) (C : Config) : Irmin.Append_only.Maker
 module Atomic_write_ext (IO : IO) (C : Config) : Irmin.Atomic_write.Maker
-module Maker_ext (IO : IO) (Obj : Config) (Ref : Config) : Irmin.Maker
+
+module Maker_ext (IO : IO) (Obj : Config) (Ref : Config) :
+  Irmin.Maker with type info = Irmin.Info.default
 
 (** {1 In-memory IO mocks} *)