+++ src/irmin/type.ml
@@ -15,3 +15,9 @@
  *)
 
 include Repr
+
+module type Defaultable = sig
+  include S
+
+  val default : t
+end